Create README.md
A basic website to be the front end to a backend api. 

A few notes this was a introductory project. The $http calls are hardcoded to be sending to localhost:3000. 

Also the service being used Data should be exchanged with the use of session data. 

The link are also hardcoded which means for this project to work it needs to be in the root folder of your webserver. 

The number of controllers can be lessened for instance the add/edit controllers basically use the same html file named differently.